requirements.txt
appnope==0.1.0
asn1crypto==1.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
camelot-py==0.7.3
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cryptography==2.7
cycler==0.10.0
cytoolz==0.10.0
dask==2.6.0
decorator==4.4.0
et-xmlfile==1.0.1
geographiclib==1.50
geopy==1.20.0
idna==2.8
imageio==2.6.1
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
olefile==0.46
opencv-python==4.1.1.26
openpyxl==3.0.0
pandas==0.25.2
parso==0.5.1
pdfminer.six==20191020
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.9.0
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
shap==0.31.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.1.7
xgboost==0.90
xlrd==1.2.0
